% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMap_footer.R
\name{classMap_footer}
\alias{classMap_footer}
\title{Title}
\usage{
classMap_footer(sort = "公職考試", dependency = NULL)
}
\arguments{
\item{dependency}{}

\item{category}{a character. "公職考試"or"研究所"}
}
\value{

}
\description{
Title
}
\examples{
none.
}
