% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMap_smallBoard.R
\name{classMap_smallBoard}
\alias{classMap_smallBoard}
\title{Title 生成課程地圖背景小卡片}
\usage{
classMap_smallBoard(
  sort = "選修課程",
  classtype = "先修課程",
  dependency = NULL
)
}
\arguments{
\item{sort}{a character "必修課程" or "選修課程"}

\item{classtype}{a character "全部課程" or "先修課程"}

\item{dependency}{default=NULL}
}
\value{

}
\description{
Title 生成課程地圖背景小卡片
}
\examples{
none
}
