% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursemap.R
\name{coursemap}
\alias{coursemap}
\title{生成課程地圖(分類和背景)卡片}
\usage{
coursemap(sort, dependency = NULL)
}
\arguments{
\item{sort}{"研究所" or "公職考試"}

\item{dependency}{}
}
\value{

}
\description{
生成課程地圖(分類和背景)卡片
}
\examples{
none
}
