% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{.page}
\alias{.page}
\title{<Add Title>}
\usage{
.page(
  content = sectionPanel(),
  footerOn = c("首頁", "找人", "找事/物", "經濟數據"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
