% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebookcard.R
\name{ebookcard}
\alias{ebookcard}
\title{產生電子書可借存量小卡片}
\usage{
ebookcard(number, dependency = NULL)
}
\arguments{
\item{number}{a character. 顯示存量如"5/10"}

\item{dependency}{default=NULL}
}
\description{
產生電子書可借存量小卡片
}
\examples{
none.
}
