% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equipmentCard.R
\name{equipmentCard}
\alias{equipmentCard}
\title{器材借用 卡片}
\usage{
equipmentCard(
  eqpt = "電子閱讀器",
  state = "借用",
  number = NULL,
  dependency = NULL
)
}
\arguments{
\item{eqpt}{a character. "電子閱讀器"or"電腦教室"or"影音錄製室"}

\item{state}{a character. 借用or無法借用or已借用}

\item{number}{a character. 數量or無or剩餘天數}

\item{dependency}{}
}
\value{

}
\description{
器材借用 卡片
}
\examples{
none.
}
