% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findthing.R
\name{findthing}
\alias{findthing}
\title{Title}
\usage{
findthing(category = "重要消息", dependency = NULL)
}
\arguments{
\item{category}{a character. 重要消息", "講座資訊", "器材借用" or "課程地圖"}

\item{dependency}{}
}
\value{

}
\description{
Title
}
\examples{
none.
}
