% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findthingPageContent.R
\name{findthingPageContent}
\alias{findthingPageContent}
\title{Title}
\usage{
findthingPageContent(dependency = NULL)
}
\arguments{
\item{dependency}{}

\item{category}{a character. 重要消息", "講座資訊", "器材借用" or "課程地圖"}
}
\value{

}
\description{
Title
}
\examples{
none.
}
