% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer_support.R
\name{footer}
\alias{footer}
\title{Footer}
\usage{
footer(onType = c("首頁", "找人", "找事/物", "經濟數據"))
}
\arguments{
\item{onType}{one of c("首頁","找人", "找事/物","經濟數據")}
}
\value{

}
\description{
Footer
}
\examples{
footer(onType="首頁") |> econWeb::browseTag2()
}
