% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_footer.R
\name{FooterButton}
\alias{FooterButton}
\title{Generate one of the four footer btn functions}
\usage{
FooterButton(onType)
}
\arguments{
\item{onType}{one of c("首頁","找人", "找事/物","經濟數據")}
}
\value{
a list with on() and off() method and default property .tag
}
\description{
Generate one of the four footer btn functions
}
