% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importantInfo.R
\name{importantInfo}
\alias{importantInfo}
\title{Title}
\usage{
importantInfo(kind = "系友會資訊", dependency = NULL)
}
\arguments{
\item{kind}{a character. "系友會資訊" or "實習資訊"}

\item{dependency}{}
}
\value{

}
\description{
Title
}
\examples{
none.
}
