% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectureCard.R
\name{lectureCard}
\alias{lectureCard}
\title{講座卡片}
\usage{
lectureCard(
  date = "5/24",
 
    topic = "Migrants from a Different Shore: Earnings and Economic Assimilation of Immigrants from China in the United States",
  lecturer = "臺北大學經濟學系 劉曦敏",
  meet = "Y",
  dependency = NULL
)
}
\arguments{
\item{date}{a character. 日期}

\item{topic}{a character. 主題}

\item{lecturer}{a character. 講者}

\item{meet}{a character. 遠距連結有無}

\item{dependency}{}
}
\value{

}
\description{
講座卡片
}
\examples{
none.
}
