% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskbtn.R
\name{maskbtn}
\alias{maskbtn}
\title{<Add Title>}
\usage{
maskbtn(
  type = c("首頁", "找人", "找事/物", "經濟數據"),
  state = c("on", "off"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
