% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page2.R
\name{Page}
\alias{Page}
\title{initiate a page constructor}
\usage{
Page(pageTitle = "首頁")
}
\arguments{
\item{pageTitle}{}
}
\value{
a shiny tag
}
\description{
initiate a page constructor
}
\examples{
\dontrun{
# construct homepage
homepage = Page("首頁")
homepage$create()
homepage$footer$button[[1]]$on()
homepage$add_content("首頁內容")
homepage$tag() |> econWeb::browseTag2()
homepage$footer$button[[1]]$return()
homepage$tage() |> econWeb::browseTag2()

pplPage = Page("找人")
pplPage$create()
pplPage$footer$button[[2]]$on()
pplPage$add_content("找人內容")
pplPage$tag() |> econWeb::browseTag2()

thingPage = Page("找事/物")
thingPage$create()
thingPage$footer$button[[3]]$on()
thingPage$add_content("找事/物內容")
thingPage$tag() |> econWeb::browseTag2()

statPage = Page("經濟數據")
statPage$create()
statPage$footer$button[[4]]$on()
statPage$add_content("數據內容")
statPage$tag() |> econWeb::browseTag2()}
}
