% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{pages}
\alias{pages}
\title{Page template constructor}
\usage{
pages(pageTitle = "首頁", content = NULL)
}
\arguments{
\item{pageTitle}{title of the page, default at "首頁". choices are "首頁","找人", "找事/物","經濟數據"}

\item{content}{a shiny.tag to be inserted into page content.}
}
\value{

}
\description{
Page template constructor
}
\examples{
none.
}
