% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pillButton.R
\name{pillButton}
\alias{pillButton}
\title{器材借用 按鈕}
\usage{
pillButton(state = "借用", number = NULL, dependency = NULL)
}
\arguments{
\item{state}{a character. 借用or無法借用or已借用}

\item{number}{a character. 數量or無or剩餘天數}

\item{dependency}{}
}
\value{

}
\description{
器材借用 按鈕
}
\examples{
none.
}
