% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectionPanel_support.R
\name{sectionPanel}
\alias{sectionPanel}
\title{Page section builder}
\usage{
sectionPanel(
  title = "經濟數據",
  imgUrl = "https://ntpuecon.github.io/econApp/assets/img/graphdemo.png",
  imgWidget = NULL,
  storyTitle = NULL,
  story = NULL,
  dependency = NULL
)
}
\arguments{
\item{dependency}{}
}
\value{

}
\description{
Page section builder
}
\examples{
none
}
