% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePillButton.R
\name{simplePillButton}
\alias{simplePillButton}
\title{器材借用 按鈕, 無數量}
\usage{
simplePillButton(state = "借用", dependency = NULL)
}
\arguments{
\item{state}{a character. 借用or無法借用}

\item{dependency}{}
}
\value{

}
\description{
器材借用 按鈕, 無數量
}
\examples{
none.
}
