% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statCard.R
\name{statCard}
\alias{statCard}
\title{經濟數據小卡片}
\usage{
statCard(
  title = "景氣信號",
  content = businessCycleLight(),
  color = "pink",
  dependency = NULL
)
}
\arguments{
\item{title}{a character}

\item{content}{a shiny.tag.}

\item{color}{"pink", "blue", "yellow" or "green"}

\item{dependency}{default=NULL}
}
\description{
經濟數據小卡片
}
\examples{
none.
}
