% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{useMaterialise}
\alias{useMaterialise}
\title{Attach materialise, jquery, google font icon and mobile dependencies}
\usage{
useMaterialise(tag)
}
\arguments{
\item{tag}{a tag}
}
\value{
a tag
}
\description{
Attach materialise, jquery, google font icon and mobile dependencies
}
