% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMap_classes.R
\name{classes}
\alias{classes}
\title{Title 生成課程小卡片}
\usage{
classes(classname = "總體經濟學", pin = "有", back = "初始", dependency = NULL)
}
\arguments{
\item{classname}{a character 課程名稱}

\item{pin}{"有" or "無"}

\item{back}{"初始" or "選定" or "先修課"}

\item{dependency}{default="NULL"}
}
\value{

}
\description{
Title 生成課程小卡片
}
\examples{
none
}
