% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_support.R
\name{page}
\alias{page}
\title{Page}
\usage{
page(
  sectionPanel = sectionPanel(),
  footerOn = c("首頁", "找人", "找事/物", "經濟數據")
)
}
\arguments{
\item{footerOn}{}
}
\value{

}
\description{
Page
}
\examples{
none
}
