% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{Page}
\alias{Page}
\title{Page template constructor}
\usage{
Page(pageTitle = "扶뒬매", content = NULL)
}
\arguments{
\item{pageTitle}{title of the page, default at "扶뒬매"}

\item{content}{a shiny.tag to be inserted into page content.}
}
\value{

}
\description{
Page template constructor
}
\examples{
none.
}
