% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskbtn_support.R
\name{roundMaskBtn}
\alias{roundMaskBtn}
\title{Rounded masked button}
\usage{
roundMaskBtn(
  type = c("首頁", "找人", "找事/物", "經濟數據"),
  state = c("on", "off")
)
}
\arguments{
\item{type}{one of c("首頁","找人", "找事/物","經濟數據")}

\item{state}{one of c("on", "off")}
}
\value{

}
\description{
Rounded masked button
}
\examples{
roundMaskBtn(type="首頁", state="off") |> econWeb::browseTag2()
}
