% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectionPanel.R
\name{sectionPanel}
\alias{sectionPanel}
\title{<Add Title>}
\usage{
sectionPanel(
  title = "經濟數據",
  imgUrl = "/lib/img-1/graphdemo.png",
  imgWidget = NULL,
  storyTitle = NULL,
  story = NULL,
  dependency = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
